Type typo.
svn:sync-xref-src-repo-rev: r108019

+++ src/VBox/VMM/VMMAll/IEMAllInstructions.cpp.h
@@ -1,4 +1,4 @@
-/* $Id: IEMAllInstructions.cpp.h 61660 2016-06-10 15:19:48Z knut.osmundsen@oracle.com $ */
+/* $Id: IEMAllInstructions.cpp.h 61665 2016-06-10 16:25:40Z michal.necasek@oracle.com $ */
 /** @file
  * IEM - Instruction Decoding and Emulation.
  */
@@ -1510,7 +1510,7 @@ FNIEMOP_STUB(iemOp_movups_Vps_Wps__movupd_Vpd_Wpd__movss_Vss_Wss__movsd_Vsd_Wsd)
 FNIEMOP_DEF(iemOp_movups_Wps_Vps__movupd_Wpd_Vpd__movss_Wss_Vss__movsd_Vsd_Wsd)
 {
     /* Quick hack. Need to restructure all of this later some time. */
-    uint8_t const fRelevantPrefix = pIemCpu->fPrefixes & (IEM_OP_PRF_SIZE_OP | IEM_OP_PRF_REPNZ | IEM_OP_PRF_REPZ);
+    uint32_t const fRelevantPrefix = pIemCpu->fPrefixes & (IEM_OP_PRF_SIZE_OP | IEM_OP_PRF_REPNZ | IEM_OP_PRF_REPZ);
     if (fRelevantPrefix == 0)
     {
         IEMOP_MNEMONIC("movups Wps,Vps");
